resolution d'un bug

+++ Meteo/src/main/java/com/mycompany/meteo/MeteoFrame.java
@@ -183,13 +183,13 @@ public class MeteoFrame extends JFrame {
                 JSONObject forcast = (JSONObject) get(); //on recupere tout l'objet JSon
                 JSONArray data = (JSONArray) forcast.get("data");// on recupere la premiere cle data qui contient un tableau
                 JSONObject firstEntry = (JSONObject) data.get(0); //dans data on recupere la cle 0 qui est un objet 
-                Double temperatD = (Double) firstEntry.get("temp"); //et dans la cle 0 on recupere enfin la cle temperature
+                Long temperatD = (Long) firstEntry.get("temp"); //et dans la cle 0 on recupere enfin la cle temperature
                 int temperati =temperatD.intValue();
                 String timeZone =(String)forcast.get("timezone");
                 Long time = (Long) firstEntry.get("ts");// on recupere le nombre de seconde ecoulées depuis 1970
                 
                 String tempsActuel =convertir(time,timeZone);
-                Double vent =(Double)firstEntry.get("wind_spd");
+                Double vent =(Double) firstEntry.get("wind_spd");
                 System.out.println("temps" +time);
                 timeLabel.setText("Il est " +tempsActuel+ "et la temperature est ");
                 temperature.setText(String.valueOf(temperati)+"°");
